# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API server for Minions.",
        "title": "Minions API",
        "contact": {
            "name": "Dima",
            "url": "http://t.me/BelozerovD"
        }
    },
    "paths": {
...
        "/api/stat/top-green-links-all-time": {
            "get": {
                "description": "Returns a pie chart of the top 5 green (safe) links accessed all time.",
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "Statistics"
                ],
                "summary": "Top 5 Green Links All Time",
                "responses": {
                    "200": {
                        "description": "HTML with embedded chart",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/stat/top-green-links-day": {
            "get": {
                "description": "Displays a pie chart of the top 5 green (safe) links accessed today",
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "Statistics"
                ],
                "summary": "Top 5 green links for today",
                "responses": {
                    "200": {
                        "description": "HTML with embedded chart",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Status Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/stat/top-green-links-month": {
            "get": {
                "description": "Displays a pie chart of the top 5 green (safe) links accessed this month",
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "Statistics"
                ],
                "summary": "Top 5 green links for this month",
                "responses": {
                    "200": {
                        "description": "HTML with embedded chart",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Status Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/stat/top-green-links-week": {
            "get": {
                "description": "Displays a pie chart of the top 5 green (safe) links accessed this week",
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "Statistics"
                ],
                "summary": "Top 5 green links for this week",
                "responses": {
                    "200": {
                        "description": "HTML with embedded chart",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Status Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/stat/top-red-links-all-time": {
            "get": {
                "description": "Returns a pie chart of the top 5 red (malicious) links accessed all time.",
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "Statistics"
                ],
                "summary": "Top 5 Red Links All Time",
                "responses": {
                    "200": {
                        "description": "HTML with embedded chart",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/stat/top-red-links-day": {
            "get": {
                "description": "Displays a pie chart of the top 5 red (malicious) links accessed today",
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "Statistics"
                ],
                "summary": "Top 5 red links for today",
                "responses": {
                    "200": {
                        "description": "HTML with embedded chart",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Status Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/stat/top-red-links-month": {
            "get": {
                "description": "Displays a pie chart of the top 5 red (malicious) links accessed this month",
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "Statistics"
                ],
                "summary": "Top 5 red links for this month",
                "responses": {
                    "200": {
                        "description": "HTML with embedded chart",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Status Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/stat/top-red-links-week": {
            "get": {
                "description": "Displays a pie chart of the top 5 red (malicious) links accessed this week",
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "Statistics"
                ],
                "summary": "Top 5 red links for this week",
                "responses": {
                    "200": {
                        "description": "HTML with embedded chart",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Status Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "common.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "Сообщение об ошибке",
                    "type": "string"
                }
            }
        },
...
    }
}